refactor: ViewModelProtocol 파일 분리
